bin: extended hosted-engine capabilities
added --vm-shutdown for gracefully shutdown the VM
modified --vm-stop for stopping the VM calling destroy,
renamed to --vm-poweroff
added --connect-storage for connecting the storage domain
added --start-pool for starting the storage pool

If the storage is not connected and the pool is down
the VM will not start due to missing volume.
Then you have to destroy the VM
and start the pool before starting again the VM.

--connect-storage and --start-pool are separated
allowing to run the VM without the pool if VDSM
supports it.

Change-Id: Iea0755188efe0dfe1f61da762d3b0f290b8efd2c
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>