packaging: setup: support OVF import
At customization stage allow to select disk as boot
media.
With disk boot media will be asked a path to a valid OVF
archive. The archive will be validated and needed VM
configuration parameters will be retrieved from the OVF
XML file.
The VM configuration will be validated and the image will
be imported.

Change-Id: I04b1de72c19e558b20d46e9fe6a435c594b0005d
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>