packaging: setup: handle bond and bridge interfaces
Previously all existing bond and bridge interfaces where
listed as possible nics for creating the management bridge.
If a bond interface was specified the bridge creation failed.

Now the bridge are not listed anymore and bonds are listed
only if they have slave interfaces.
Enslaved interfaces and interfaces used by other bridges
are not listed anymore as available for management bridge
creation.

If a bond interface with slaves is specified now should work
correctly.

Change-Id: If19dfa2c2b9b02ba8d997ad1758614fad3470eb5
Bug-Url: https://bugzilla.redhat.com/999577
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>