vdsm: use devices instead of drives in vm.conf
Previously the setup used the keywords drives, cdrom,
nicModel, bridge and macAddr for configuring the VM
devices.
Such keywords seem to be obsolete and lead to having
the share:exclusive option not working correctly on drives.
Now it uses the keyword devices for configuring all the devices
and drives in order to have share:exclusive support.

Change-Id: Idbe61b2ef899320f304cd30b5b365e2a523acd49
Bug-Url: https://bugzilla.redhat.com/1008391
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>