sanlock: creates metadata file
Creates the metadata file is shared by all hosts participating in the engine
high availability cluster.  Each host is assigned a 4KiB chunk according
to its host id, with a corresponding offset of host_id*4KiB.  The
initial 4KiB of the file are reserved for shared state such as a global
engine ha maintenance bit.

Change-Id: I8862209fe2fd783aaeab04d2c655afd9e1a10396
Bug-Url: https://bugzilla.redhat.com/1008501
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>