packaging: setup: check for stale VM before creation
Added a check for stale Hosted Engine VM before creating it.
In normal execution flow, this should never happen.
However, it is possible that the user starts again the VM
manually after having installed the OS and forget to poweroff
(destroy) the VM.
Now the user will be warned to properly poweroff the VM
started manually in order to continue the setup process.
The poweroff is not issued automatically because the VM can
be in a stale state after a failure and the user must be
aware that the VM is still around in order to debug the issue.

Bug-Url: https://bugzilla.redhat.com/1010980
Change-Id: I8d5dfc17e10e677959e903f03515a124d5e97611
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>