packaging: setup: handle bridge on additional hosts
Previously on an additional host hosted-engine-setup checked
for ip addr of the bridge instead of the ip addr of any existing
device. But on additional host the bridge creation is done by
the manager when adding the host.

Enabled the NIC selection for bridge creation also on additional
hosts if the bridge is not found and use its ip

Bug-Url: https://bugzilla.redhat.com/1025253
Change-Id: I81a5edada72d397c3ce30410d903840b1392034e
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>