Allow user to reboot VM from within the installer
During setup, the engine VM is created twice: once to install the base
operating system and once to install the engine.  During either of these
phases it may sometimes be necessary to reboot the VM additional times
(ie. to correct PXE boot configuration or restart the installer).
Provide an interactive means of rebooting the VM without forcing the
user to resort to external invocations of hosted-engine --vm-start, etc.

Signed-off-by: Adam Litke <alitke@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1055497
Change-Id: I176eb58b6b5400df7f98a4fa3e4ae55c2959fb93