Move VM operations into a mixin for reusability
The plugin in runvm.py has lots of useful logic for manipulating the
hosted engine VM.  This logic can be very useful to other plugins (eg.
health.py) that may want to offer the user assistance with managing the
VM.  This patch moves the VM logic into a separate class that can be
mixed in with any other plugin which may want the functionality.

Signed-off-by: Adam Litke <alitke@redhat.com>
Related-To: https://bugzilla.redhat.com/1034788
Change-Id: I811f6712cb5c9324c6a3acca0bfc5bc9195b25ef