packaging: setup: skip space check on additional hosts
On first host we need to check if we have enough space too.
We must skip this check on additional hosts because the space is
already filled with the Hosted Engine VM image.
Sadly we can't go back to previous customization stage so here
we can only fail the setup.

Change-Id: Ie6dae85998d8f946cca6d1f8ba3750353af002e1
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>