packaging: setup: always wait for host addition
Without this patch, we wait only on first host and not on additional.

This causes (among other things?) HA to start before host-deploy.
At that point, the vdsm cert has the hostname as CN. host-deploy then
creates it again with the IP address as CN (caused by HE deploy, a
different matter).

Prompt for required networks if needed.

Change-Id: Ib8537b25493fa0263d09887b04935796d601c6eb
Bug-Url: https://bugzilla.redhat.com/1178535
Signed-off-by: Yedidyah Bar David <didi@redhat.com>