packaging: setup: asking about retriyng VG creation with force option
On dirty storage devices CreateVG could fail, force option can
prevent failures destroying any existent data.
If VDSM detect that the LUN is alredy used it ask to the user
if he want to use the force option to explicitly clean it
up.
If, for any reasons, createVG still fails, it asks again
to the user if he want to retry with force option to
clean it or if he prefer to abort.

Change-Id: I496c34e6b9f0d84443a8d5bc68d77916be6cb504
Bug-Url: https://bugzilla.redhat.com/1142098
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>