packaging: setup: handling previously generated answerfile with wrong types
Till 3.5 we were threating some VM parameters as integer
is read from the OVF file and as string if manually entered.
It was not really an issue cause we were than generating
a temporary configuration fail for vdsClient usage with an
implicit string conversion.
Now we are using vdsclient and so we have to use the correct
type and so we need also to handle answerfile generated from
previously release where the type could be not correct.

Change-Id: I07c296c4ef92430f03a5b23a2a996db9c643a9c5
Bug-Url: https://bugzilla.redhat.com/1198725
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>