Adapt to VDSM 3.7 StorageDomainInfo API
In VDSM 3.7 StorageDomainInfo returns a string for
the "type" key in the result dictionary.
The string should be ones between the names of the constants in
the enum type StorageDomainType defined in vdsm/rpc/vdsmapi-schema.json

The VDSM schema is missing GLUSTERFS but it's using it anyway.

Change-Id: Id5f45d088c943c62e0e5761cca1a565616b2dcc5
See-Also: https://bugzilla.redhat.com/1215039
Bug-Url: https://bugzilla.redhat.com/1213307
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>