packaging: setup: remove connectStoragePool on iSCSI additional hosts
On hosted-engine from oVirt 3.6 the hosted-engine storage domain is
not anymore connected to the bootstrap storage pool so the spUUID
is now blanck on additional hosts.
Removing initial connectStoragePool on additional iSCSI/FC hosts,
adding a getStorageDomainStats call to rely on a side effect of
that and get /rhev/data-center/... populated.

Change-Id: I3bec0e6ad5489a90ebfad198f7934cf776562f6a
Bug-Url: https://bugzilla.redhat.com/1247942
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>