packaging: setup: disallow redeploy on dirty storage
In the past we had the capability to re-deploy hosted-engine
over a previously partially deployed system; the trick was simply to
let the user boot the VM from CD and than exit without touching
the engine VM disk and than continue frm there.
With the appliance it doesn't make sense cause it will always
overwrite the previously deployed one destroying it.
Explicitly disable it to prevent confusion.

Change-Id: I3a2a1ab7b74450ad9341bd6d3e7694d99c97f5da
Bug-Url: https://bugzilla.redhat.com/1238709
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>