packaging: setup: proposing socat instead of virsh
In the past we proposed to connect to the VM via remote-viewer
or virsh for who just want to use a text only client.
virsh requires authentication as vdsm@ovirt user, but
the console is also available over a unix domain socket.
Proposing the user to use socat to connect to that.

Bug-Url: https://bugzilla.redhat.com/1215436

Change-Id: I5abc8ce5ff048b96a456fdecdb300848e063383c
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>