packaging: setup: avoid asking about mgmt bridge nic on additional hosts
On additional host we are not directly creating the management bridge.
We simply call host.add via REST API and the engine will do the rest
via host-deploy. So avoid asking about the interface to create the
management bridge on cause we are not honoring it.
The engine will choose it exactly as it does for regular non HE hosts.

Bug-Url: https://bugzilla.redhat.com/1269430

Change-Id: Ibcb56997b17fe04dbccf25fa2a3bd9c79e5da8e9
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>