setup: ask host address on additional hosts
Use hostname.getHostname from ovirt_setup_lib
to ask and validate the host address on additional
host.

On the first host we already ask the management
interface to create the management bridge on
and so we can only take the address that uniquely
resolves there.
On additional hosts we simply rely on host-deploy
to create the management bridge on the interface
where the address supplied to hosts.add resolves on
and so it's worth to ask the user with a reasonable
default.

Change-Id: Ic5ce041e47fef02d60aa46159ff56f9e0622d5fc
Bug-Url: https://bugzilla.redhat.com/1251968
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>