packaging: setup: fixing _storagePoolConnection on additional hosts
Before 3.6.1 _storagePoolConnection was not required on additional
hosts since the hosted-engine storage domain wasn't connected
to any storage pool.
With 3.6.1 the engine is able to import the hosted-engine
storage domain and attach it to its datacenter storagePool.
So _storagePoolConnection is needed again on additional hosts
and unfortunately we have to fix it.

Change-Id: I44389c06c2d2611bcead7b250f2d9f262d2ffce8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1291234
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>
(cherry picked from commit 25b97ca5b2682449995f395782475e2548fb3a04)