storage: avoid connecting to the storage pool to check its name
Since 3.6.1 the hosted engine storage domain get imported
into the engine and attached to the storage pool of the
data-center that contains the hosted-engine hosts.
hosted-engine-setup on additional hosts was connecting the
bootstrap storage pool just to check its name but now it
cannot anymore since the master storage domain of the
datacenter storage pool could be a regular storage domain
on a storage server that hosted-engine-setup didn't
connected at all.
Avoid it and simply rely on the name in the answerfile gathered
from the first host.

Change-Id: I912eb3bdd69567e1902ab28f71ce16496b4e9bb3
Bug-Url: https://bugzilla.redhat.com/1293892
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>