storage: resetting spUUID to a random value only if really needed
hosted-engine-setup generates answer file with a blank SP UUID
for additional hosts cause it correctly disconnected its SD
from the boostrap storage pool.
If we want to use the answerfile for not-interactively replaing
the scenario we need a valid spUUID value otherwise we are not
able to create the bootstrap storage pool.
On the other side, passing an answerfile via CLI is a valid
alternative instead of download it from another host via scp.
And in that case we should honour the BLANK UUID.
So resetting spUUID to blank only if really needed to create
a new boostrap storage pool.

Change-Id: I67569bc6902a176565269f2bcbec569b51549181
Bug-Url: https://bugzilla.redhat.com/1277010
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>