networking: enforcing unique match between hostname and mgmt interface
hosted-engine-setup has to create the management bridge before
starting the engine VM otherwise it will have no connectivity.
Then it will trigger host-deploy calling hosts.add on the engine API
passing just an address parameter to identify the host;
the address should be the host name (cfg rhbz#1178535)
On host-deploy, VDSM will try to re-create the management bridge on the
interface where address resolves to. This can randomly lead to a wrong
network configuration if address can be resolved to more than one interface.
Enforcing unique match between address and the IP of the management
interface to avoid this.
This is compatible with the current version of oVirt node since it let the user
to configure just one interface and so there is no ambiguity by construction there
but it should be re.tested with node.next.

Change-Id: I478b499c6e1b3a6160822f4c58966c6d29f4e42d
Bug-Url: https://bugzilla.redhat.com/1291083
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>