storage: honor SP=BLANK if provided on CLI answerfile
When hosted-engine-setup deploys an additional host
it reads back the SP UUID from disk; this could be BLANK
(just a after the setup of the first host) or a valid value
if the engine already auto-imported the hosted-engine
storage domain.

Deploying additional host, hosted-engine-setup parses the
answerfile from the first host which contains SP=BLANK.
When the user runs hosted-engine-setup interactively
it parses the first host-answerfile after checking the
storage so hosted-engine-setup will always continue
with SP=BLANK.
Instead, if the user wants to run it unattended, he can
append an answerfile on CLI. Also in that case we have to
honor SP=BLANK if present (if not it's just a random value).

Raising if prepareImage fails since continuing will just
hide the real issue.

Change-Id: I56d2a671db4f288bcb72218a88cb071e241ff781
Bug-Url: https://bugzilla.redhat.com/1311317
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>