upgrade: prevent deploying additional 3.6 host on a 3.5 setup
Deploying an additional 3.6 hosted-engine host requires an
existing 3.6 system. Mixing 3.6 and 3.5 hosts in the same HE
cluster is not supported: in order to deploy a new 3.6 host,
the user has to update existing hosts to 3.6,
and only then add the new host.
Looking for the presence of the configuration image uuid
in the answerfile, we can detect if the answerfile is
from 3.5 or 3.6, and warn the user if needed.

Change-Id: I9283e5d45eba63b4bfbb99343af1b4331856b6a5
Bug-Url: https://bugzilla.redhat.com/1306825
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>