tool: let the user pass a custom vm.conf
Since 3.6.z the user will be able to edit the engine VM
configuration from the engine. This could lead to an issue
if the new configuration will prevent the engine VM to start
since in order to edit it again the user needs the engine.
Passing an alternative vm.conf from CLI the user could be
able to temporary amend what was wrong in the master
configuration on the shared storage in order to be able to
start again the engine VM and definitively fix it.

Change-Id: Id0bd98648600872ea8145c1ad90cab3264b430a9
Bug-Url: https://bugzilla.redhat.com/1309572
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>