network: skip NetworkManager check/disable via answerfile
Adding the capability to skip NM check and disabling by
host-deploy via answerfile (dafault enforcing).
In the normal flow, host-deploy will disable NetworkManager if active.
When host-deploy runs, the management bridge has already been created
and the engine VM is already running so disabling NetworkManager on
the fly can be risky for the engine VM networking.
To avoid that, ovirt-hosted-engine-setup will check NetworkManager
status and eventually ask the user to stop:
this behavior wasn't configurable.
On the other side Next Generation Node is based on Cockpit, which
requires NetworkManager to report the network status.
Next Generation Node comes with NetworkManager-config-server
so NetworkManager there should be passive enough.
With this patch we make its check optionally skippable via answerfile.

Change-Id: Iaec165cfb8bf93a3a17af20b23ce97ec2ad3223d
Bug-Url: https://bugzilla.redhat.com/1337050
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>