Add --check-deployed to hosted-engine
hosted-engine already checks whether or not vmid exists as a proxy
to find whether the engine is deployed.

Add a trivial function which checks this, so other utilities can
verify this without parsing the output of other commands to see if
a bad exit code is from exit_not_deployed or for another reason.

Change-Id: I5c876ccccccb79df858538438e6767fec8ed32ed
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1324921
Signed-off-by: Ryan Barry <rbarry@redhat.com>