storage: pass mount options via answerfile
Some scenario could require custom mount options calling
connectStorageServer on the hosted-engine storage domain.
Adding an optional parameter to store it in the conf file
if a value has been supplied via answerfile.
Doing it interactively will require to propose a
reasonable default and validate the user input so it
will addressed in a future patch.

Change-Id: I8fb6e6ff4cfd9b794c54b6a4c8e9de3ee2de9289
Bug-Url: https://bugzilla.redhat.com/1298693
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>