upgrade: migrate file based lockspace and metadata
On oVirt 3.4, metadata and lockspace area were based
on files directly saved in the NFS storage.
This would not work with iSCSI and so with 3.5 we
moved to regular VDSM volumes also on NFS. This
is a procedure to upgrade system started with 3.4
to use VDSM volumes.
The upgrade procedure will scan for the metadata
and the lockspace area, if they aren't on the
storage domain it will trigger the volume creation.
Volume creation will also remove the previous file
and it will replace it with a symlink pointing to the
new volume.
Upon restart, all the hosts will point to the new
volume since they'll simply consume the symlink.

Change-Id: If64a1325dc4187b61bff9eb1d2763995c8387e46
Bug-Url: https://bugzilla.redhat.com/1313917
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>