cloud-init: add a comment to explain how to avoid 1372260
Due to a kernel bug (BZ#1372260), 'systemctl restart sshd'
could be blocked forverer and this will block also our
cloud-init script and so the deployment till
hosted-engine-setup will abort at a timeout.
Restart sshd in a later section (from bootcmd to runcmd)
to try avoiding BZ#1372260 and, if this is still not enough,
start it in background to avoid getting stuck also if
sshd is not available.

Change-Id: I4f3380c02ed1b95391a2fca31d82a43c5408f228
Bug-Url: https://bugzilla.redhat.com/1416023
Bug-Url: https://bugzilla.redhat.com/1409203
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>