upgrade: fixing vm.conf patching method
For the first boot of the new appliance
we have to patch on the fly vm.conf to inject
the cloud-init seed image.
Patch https://gerrit.ovirt.org/67924 slightly changed
the way to read vm.conf, fixing according to it.
With the previous interface it was possible to refresh
from the shared storage (initial vm.conf or
the OVF_STORE if available) to a temporary file and
edit in place. Now we can just refresh vm.conf
in its final local destination under /var/run, read
from there, inject the cloud-init iso and write
to a temporary file as before.

Change-Id: Ib6fbbdeb5bcb84d59982c12ebdc71fb20c082b29
Bug-Url: https://bugzilla.redhat.com/1420283
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>