packaging: fixing bad permissions on answer file from the past
answers.conf is in the configuration volume on the shared
storage since 3.6 while it was just a local file before.
The 3.5 -> 3.6 upgrade procedure is able to move it to
the shared storage but it runs in ovirt-ha-agent deamon
as vdsm user so answers.conf should be readable by vdsm user.

Bug-Url: https://bugzilla.redhat.com/1422962

Change-Id: If234bae54f204f733016e4268f18ec2c86802aca
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>