gluster: query mount options only if on gluster
The query for additional mount options has been hidden
with https://gerrit.ovirt.org/69441/ since the engine
is in general not ready to handle it.
GlusterFS is an exception because only in that case the
engine is able to handle it.
gdeploy was already passing the value
via answerfile but, in the case of a manual HE deployment
on a system where gdeploy just created the volumes,
there is no interactive question.
Adding it back only for gluster.

Change-Id: I8ddd9620119c9918f9a36beb2bcb86729d825ac4
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1426517
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>