storage: call disconnectStorageServer on the bootstrap SD
The bootstrap SD is created just to be used as the master
storage domain to boostrap the hosted-engine VM using
a temporary loopback mounted file.
We were removing the loopback device and removing its file
but we weren't calling disconnectStorageServer to
cleanly unmout it resulting in a temporary (till
reboot) leftover mount point.

Change-Id: Ida4469a9059bc17d3af821f6003ad9dfb28aad43
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1449565
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>