upgrade: enforce metadata area compatibility level
Ensure that the all the host entries in the metadata area are in
a compatible shape or force the user to remove decommissioned
hosts from there.
We were already checking compatibility level against the engine
for exisitng hosts but the matadata area could still contain
entries for older decomisisoned hosts.

Change-Id: I5fccc3d01f1582c26b03713e1bbce23aa7ab074b
Bug-Url: https://bugzilla.redhat.com/1481680
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>