WIP - Provide an Ansible installation script
This script does the following:
 1. Downloads all the required packages
 2. Unpacks the engine appliance image
 3. Creates cloud init files
 4. Starts a local vm with the engine appliance and the cloud init files
 5. Setups the engine (as part of the previous step)
 6. Adds the host as a managed host to the locally running vm
 7. Adds storage domain and disks
 8. Adds a vm that will later become the hosted engine vm in the engine
 9. Destroys the local vm
 10. Creates the hosted engine configuration files using Ansible templates
 11. Copies the conf files to the relevant location on host and on storage
 12. Copies the local vm disk to the storage
 13. Starts and enables the agent and the broker

This is a work in progress, this script takes a long time to run (mainly due to disk creation, but also due to some
other workarounds), it doesn't support other storage options than nfs, and it contains some workarounds that
need to be fixed.

For now, the vars file need to be manually edited and filled with the desired values.

Change-Id: I979aec3ad4f1da789b1aeeb90ca00df6cc2507fc
Signed-off-by: Jenny Tokar <jtokar@redhat.com>