iscsi: multipath: connect the whole portal group
Connect all the portals of a portal group to enable
iSCSI multipath.
To be as backward compatibile as possible and avoid
requiring changes in host-deploy, it saves the
multiple IP and port values as string composed
by multiple comma separated values with the
same key used by previous version.
It assumes that the login is the same for all
the portals.

Change-Id: I4cc9b5a7f281c87fb4deb7ed8ce51e35b3c10926
Requires: https://gerrit.ovirt.org/82094
Bug-Url: https://bugzilla.redhat.com/1149579
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>