ansible: network: always inject into /etc/hosts
At host-deploy time, the engine VM is going to temporary loose
external network connectivity since host-deploy will temporary
break natting rules and so name resolution, if based
on an external DNS, is going to fail as well.
Always inject an entry for the host into /etc/hosts on the engine
VM and eventually just remove it at the end, if not needed
on the target VM, to prevent that.

Change-Id: I970b4b5212978ec54ddde65df9231a7e71a4d14e
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>