ansible: temporary disable reverse DNS lookups
The engine VM is started on the default natted libvirt VM.
DNS service is provided by dnsmasq on the host but
host-deploy/vdsm disrupts it configuring the
management network on the host and so
ssh logins performed by ansible becomes really slow.
Temporary disable sshd reverse DNS lookups via cloud-init
and restore it via ansible at the end as a workaround.

Change-Id: Ic17ffb6d39bac0a0f1b398e269f0ae7d3d247e53
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>