ansible: fix configration volume templates
We used to save the answer file generated on the first
host in order to consume it, at 3.6 time, as an answer file
to run hosted-engine-setup on additional hosts to
add them to the env.
Now we can safely do it from the engine.
The only flow that still access it is
--upgrade-appliance which only supports
3.6/el6 -> 4.0/el7 upgrades.
Altought the answrfile is not really used, is still
worth to fill it with coherent values.

Minor fixes in version.j2 and hosted-engine.conf.j2

Change-Id: I6642a53cec2f25b0b1d560201bc93b44289bc6ba
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>