ansible: remove cloud-init generation question
In the ansible based flow we are going to
execute engine-setup on the engine VM
via ansible and not via cloud-init.
The cloud-init initial file for the local
bootstrap VM should also address different
configurations so its cloud-init image will
be not be backward compatible between the two flows.
Remove the question on the ansible flow.

Change-Id: I796fe645cc82a677844fd0698fdbc61ee9f6c6d1
Bug-Url: https://bugzilla.redhat.com/1532213
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>