vintage: iscsi: avoid failing the setup if just one path is down
Avoid failing the setup if it received a multipath
configuration via answerfile and just one of the path
there is down.
Increase VM_LIVELINESS_CHECK_TIMEOUT to 900
seconds since vdsm doesn't connect storage path
in parallel so the detection of multiple failed
storage paths could take a significative
amount of time.
With this fix hosted-engine-setup is able
to conclude the deployemnt if at least a
storage path is working but the engine
will still completely ignore broken paths
at auto-import time: adding an explicit
warning.

Change-Id: Ibdf04908f1a3f3c8437e6b1c713cc34e5181fbc5
Bug-Url: https://bugzilla.redhat.com/1543424
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>