Ansible: Check available space on storage domain before adding it to the cluster
Add the storage domain in a 'present' state is a mistake because then it
cannot be removed if is not suitable, since it is the only storage domain.
Thus, adding the storage domain in a 'unattached' state which then can be
remove if needed, even if it is the only storage domain.

Change-Id: I035136043b4eff54da022694fe78dce36e178b85
Bug-url: https://bugzilla.redhat.com/1560610
Signed-off-by: Ido Rosenzwig <irosenzw@redhat.com>