Remove engine FQDN temporary tweak at the end
The bootstrap engine VM temporary runs with its
final FQDN over libvirt default natted network
with a different local address.
SSO internal module requires to resolve engine
FQDN and so we are tweaking it via an entry in
/etc/hosts.
Forcing the engine FQDN to resolve on the
address on the natted network at boostrap time.
Removing it at the end.

Change-Id: I41958f581b300f5ff41edb0138d1c74a5464317d
Bug-Url: https://bugzilla.redhat.com/1576310
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>