console: open serial console via socat
HE VM serial console device as set by the
engine (with patch
http://gerrit.ovirt.org/92864/ )
are over a unix domain socket instead of
plain pty so virsh console doesn't work with
that.
Using socat instead.
If the serial console device doesn't appear
in the output of virsh -r dumpxml, it will
try to fallback to virsh console for
compatibility with engine < 4.2.

Change-Id: I7fb2abb60a08b9e83f89609f959223be4a65ce59
Bug-Url: https://bugzilla.redhat.com/1590943
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>