checks: enforce firewalld status
Ensure that firewalld is running and
not masked or fail early with a clean
error message.
Prefering systemd module over service_facts
since the latter doesn't explicitly reports
masked status while the former does.
Masking a running service doesn't implicitly
stop it and so we need systemd module to check
both the condition.

Change-Id: I956791444bbbccefa34b435f9253dbc12447f33f
Bug-Url: https://bugzilla.redhat.com/1608467
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>