upgrade: force firewalld to use individual calls
hosted-engine-setup still contains the code
to let the user perform a clean upgrade
of the engine VM from 3.6/el6 to 4.0/el7.
This upgrade is based on an old engine VM
appliance with a 4.0 engine (we cannot skip
versions on engine side) and that old
appliance contains a firewalld version
still affected by rhbz#1446162
Force individual calls in firewalld
configuration from the cloud-init script
used for the upgrade to workaround that bug.
Reset initial firewalld configuration
after engine-setup to avoid impacts
on VM performances.

Change-Id: I18cae02d3841247a9888875163e531dfd7dd229f
Bug-Url: https://bugzilla.redhat.com/1621015
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>