restore: temporary "disable" host fencing
When the user tries to restore from a backup,
the engine can try to fence an host
that was configured in the restored backup
but its not available right now.
So also the host used for the restore process
could potentially be fenced in the middle
of the restore if redeployed and so not
operational.
Temporary disabling host fencing when deploying
restoring a backup on the fly by setting a
huge value for DisableFenceAtStartupInSec.
Its original value will be restored at the end.

Change-Id: Ib711f2ed077e5c9216cdf94e7a05ed2f7a64d41a
Bug-Url: https://bugzilla.redhat.com/1640097
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>