restore: update vm_dynamic for VMs on redeploy host
We can only redeploy on an host with no
running VMs (we eventually have also to deactivate
vdsm and reconfigure libvirt) and in order to reuse
an host that was already part of the environment
we have also to filter it out at restore time.
But in order to do that we have to remove
entries on vm_dynamic pointing to that host
(run_on_vds, migrating_to_vds).
Let's set status=Down for that,
the engine will sync their real status
during its grace period.

Change-Id: I6b79de26e5e9f7ca643d44e74c587e7aef9fb946
Bug-Url: https://bugzilla.redhat.com/1645757
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>