Set OvfUpdateIntervalInMinutes config to 1 during deploy
The deploy script waits until the OVF file is written
to the storage domain. Sometimes, probably because of a race,
the OVF write is not triggered by update VM and
the deploy script would have to wait for 60 minutes.

This patch changes the setting during the deploy,
so the OVF is stored every minute.

On master the same change has been introduced with:
https://github.com/oVirt/ovirt-ansible-hosted-engine-setup/pull/51
on the ansible role consumed there.

Change-Id: Ibd46f5b3ff505e78a5adce72c096f5a51b3e1113
Bug-Url: https://bugzilla.redhat.com/1644748
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>