stig: prompt for openscap security profile
Ask the user if they would like to apply a default security profile for
the hosted engine VM.  The security profiles are selected according from
the host's distribution and scap-security-guide content, which will be
changed in the future to a customized security profile for the hosted
engine.  Applying security profiles are disabled by default.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1392051
Change-Id: I86e2bdfcbab948ae23ed72a3e38dca897db91868
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>