network: let the user interactively validate host address
Let the user interactively validate the host address
that will be used to add the host to the engine.
In the past it was using by default socket.gethostname()
which could lead to strange behaviours on multi nic
environments where the default hostname is not correctly
set.

Change-Id: Icc047d780ac88a7bde15319a7100ddc698f8d36f
Bug-Url: https://bugzilla.redhat.com/1692460
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>
(cherry picked from commit 237ed04f4a243d8ffc782b92cabef83733714489)