Convert appliance size to float
If appliance size as parsed from ovf is not an integer, but a float,
int(size_as_string) fails. Convert it first to a float, then to int.

Without this patch:

[ ERROR ] invalid literal for int() with base 10: '51.00195312406868'

Seems like it started happening by moving creation of the appliance on
el8 and using python3:

https://gerrit.ovirt.org/104272

Also do the same when parsing the results from ansible create_local_vm
in variable otopi_appliance_disk_size. Without this, the generated disk
is smaller than the appliance one, and qemu-img refuses to copy, failing
with "output file is smaller than input file".

Change-Id: I582c14aafc36996971aeceac5929a081fd65e1db
Signed-off-by: Yedidyah Bar David <didi@redhat.com>