cleanup: Remove more stuff
- OVN

- All networks except for ovirtmgmt

Without latter, the following flow did not work as expected:

I wanted to reproduce bug 1893385. To do this, I manually created a
dummy network interface, configured an oVirt network and attached it to
it, then did backup/restore. As expected, it failed on restore. So I
manually created the dummy interface also on the restored machine and
tried again, and it worked. Then I ran the cleanup script and tried
again, and it didn't fail - vdsm still had the dummy network/nic in its
cache.

Change-Id: I4753ee19e5f2efe9cbb54ffdc96c7a6f61170f7b
Signed-off-by: Yedidyah Bar David <didi@redhat.com>