Consider missing 'available' as broken storage
Under certain conditions the result of 'Activate Storage Domain',
although successful, does not return the 'available' space in the
domain. Do not fail deployment in this case, but just consider the
storage broken and continue the loop prompting for storage.

I failed to reproduce this myself, thus do not add an extra 'if' with
specific text. The reported case seems to have been "loud" enough that
hopefully the user would be able to login to the admin UI, understand
that something was not well (even if it's not completely clear what the
user would have seen, other than lots of "events"), and retry.

I do add an 'if' and emit some generic text, just so that if the storage
creation playbook didn't emit any error (which might have been the
case), the user will not be completely confused as to why we ask again.

Bug-Url: https://bugzilla.redhat.com/1662657
Change-Id: I2c248c0bd8948203aa8dfd41bdeffb9dddff4a30
Signed-off-by: Yedidyah Bar David <didi@redhat.com>