prefer IPv6 address on mixed network stack
when we are checking our IP address against the supplied engine's IP we
look up both IPv4 and IPv6 in this orded, but match only the first find.
This works fine in IPv4-only or IPv6-only environments, as well as with
forceIPv4 and forceIPv6 options. But in mixed network stack by default
this breaks since DNS resolution prefers IPv6 instead.
This patch just changes the order and returns IPv6 address first,
following the way DNS is resolved.

Change-Id: I6f2c03bd32f7a6f7b73471158ec13250c6fcb8f9
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>