Move shared python constants
...to a location that is not python-version-specific.

Create a new module src/ansible/constants.py that includes the few
constants that are used by both setup code and ansible callbacks code.

This is needed since newer ansible uses python 3.8, and setup code uses
python 3.6. The alternative would have been to package this code for
both python versions. Since this is a very small amount of code used
only in few places, it does not seem worth it.

Change-Id: Ia411fcaf8a9341d4e10eebf0354868ff70c6a937
Signed-off-by: Yedidyah Bar David <didi@redhat.com>