Make python dateutil optional
Do not require it anymore.

If it's available, use it. Otherwise, use a simple but good enough
alternative for tz.tzlocal, the only thing we use from dateutil.

Code is copied from otopi. There, we wanted it for ovirt-host-deploy,
which was designed to run on remote systems without having anything
installed other than the bare minimum (available at the time, for the
then-supported OSes). Current patch is in preparation for using ansible
2.12, which uses python38, for which there is currently no dateutil
packaged in CentOS Stream 8.

Change-Id: Ic0b7ffbca296536a76afada241d00a9d2f99a844
Signed-off-by: Yedidyah Bar David <didi@redhat.com>