Fix ansible-core dependency
On EL8 we depends on ansible-core 2.12, which is using non-standard
Python version 3.8, but we have already rebuilt all our dependencies for
it so we are fine. But without additional effort around dependencies
rebuild we cannot upgrade to ansible-core 2.13, because it bumped
Python requirement to 3.9.

On EL9 we can use both ansible-core 2.12 and 2.13, because both of them
are using platform python, which is 3.9, so we have all required
dependencies available.

Signed-off-by: Martin Perina <mperina@redhat.com>